Authorization for controllers

+++ backend/orders/controllers/index.js
@@ -1,15 +1,16 @@
 const orderModel = require('../models')
 const productModel = require('../../products/models')
-const { getCurrentUser } = require('../../users/controllers')
+const userModel = require('../../users/models')
+const { currentUser } = require('../../functions/currentUser')
 
 const postOrder = async (req, res) => {
     const { products } = req.body
     let total = 0
     try {
-        /* Setting the bellow header for the 'getCurrentUser' controller
-           (see its return documentation) */
-        req.headers['return'] = true
-        const User = await getCurrentUser(req, res)
+        const User = await currentUser(req, res)
+        if (!User) {
+            return res.status(403).json({ error: "InvalidId" });
+        }
         /* Using a for..of loop in order to iterate asynchronously through
            the array and increase the total */
         for (const product of products) {
@@ -17,7 +18,10 @@ const postOrder = async (req, res) => {
             total += Product.price
         };
         const Order = await orderModel.create({ products, total, user: User._id })
-        console.log('res')
+
+        //Add the order to User
+        await userModel.findByIdAndUpdate(User._id, { $push: { orders: Order._id } })
+
         await res.status(200).json(Order)
     } catch (err) {
         res.status(400).json({ error: err.message })
@@ -25,6 +29,7 @@ const postOrder = async (req, res) => {
 }
 
 const getAllOrders = async (req, res) => {
+    //in production -> only admins
     try {
         const Orders = await orderModel.find()
         res.status(200).json(Orders)
@@ -34,6 +39,7 @@ const getAllOrders = async (req, res) => {
 }
 
 const getOneOrder = async (req, res) => {
+    //in production -> only admins
     try {
         const { id } = req.params;
         const Order = await orderModel.findById(id);
@@ -47,12 +53,28 @@ const getOneOrder = async (req, res) => {
 }
 
 const deleteOrder = async (req, res) => {
+    const { id } = req.params;
     try {
-        const { id } = req.params;
-        const Order = await orderModel.findByIdAndDelete(id);
-        if (!Order) {
+        const User = await currentUser(req, res)
+        if (!User) {
             return res.status(404).json({ error: "InvalidId" });
         }
+
+        const Order = await orderModel.findById(id)
+        if (!Order) {
+            return res.status(404).json({ error: "InvalidOrderId" });
+        }
+
+        //Check user authorization
+        if (User._id.toString() !== Order.user.toString() && User.role !== 'admin') {
+            return res.status(403).json({ error: "Forbidden" });
+        }
+
+        //Remove the order from User
+        await userModel.findByIdAndUpdate(User._id, { $pull: { orders: Order._id } })
+
+        //Delete order
+        await orderModel.findByIdAndDelete(id);
         res.status(200).json("Order deleted successfully");
     } catch (err) {
         res.status(400).json({ error: err.message });
