+++ backend/reviews/controllers/index.js
@@ -1,9 +1,17 @@
 const reviewModel = require('../models')
+const productModel = require('../../products/models')
+const { currentUser } = require('../../functions/currentUser')
 
 const postReview = async (req, res) => {
-    const { user, stars, title, body } = req.body
+    const { productId, stars, title, body } = req.body
     try {
-        const Review = await reviewModel.create({ user, stars, title, body })
+        const User = await currentUser(req, res)
+        if (!User) {
+            return res.status(404).json({ error: "InvalidId" });
+        }
+        const Review = await reviewModel.create({ user: User._id, stars, title, body })
+        //Add Review to Product
+        await productModel.findByIdAndUpdate(productId, { $push: { reviews: Review._id } })
         res.status(200).json(Review)
     } catch (err) {
         res.status(400).json({ error: err.message })
@@ -20,12 +28,27 @@ const getAllReviews = async (req, res) => {
 }
 
 const deleteReview = async (req, res) => {
+    const { id } = req.params;
+    const { productId } = req.body
     try {
-        const { id } = req.params;
-        const Review = await reviewModel.findByIdAndDelete(id);
-        if (!Review) {
+
+        const User = await currentUser(req, res)
+        if (!User) {
             return res.status(404).json({ error: "InvalidId" });
         }
+        const Review = await reviewModel.findById(id);
+        if (!Review) {
+            return res.status(404).json({ error: "InvalidReviewId" });
+        }
+        //Check authorization
+        if (User._id.toString() !== Review.user.toString() && User.role !== 'admin') {
+            return res.status(403).json({ error: "Forbidden" });
+        }
+
+        //Remove Review from Product
+        await productModel.findByIdAndUpdate(productId, { $pull: { reviews: Review._id } })
+
+        await reviewModel.findByIdAndDelete(id);
         res.status(200).json("Review deleted successfully");
     } catch (err) {
         res.status(400).json({ error: err.message });
