+++ backend/users/controllers/index.js
@@ -1,19 +1,7 @@
 const userModel = require('../models')
 const jwt = require('jsonwebtoken')
-
-/*  The access token is sent in the cookies on log in, it expires every 5 minutes or
-    when the user signs out, and it's refreshed with the refresh token through the 
-    'verifyAccessToken' middleware  */
-const createAccessToken = (id) => {
-    return jwt.sign({ id }, process.env.JWT_ACCESS_SECRET, { expiresIn: '5m' })
-}
-/*  The refresh token is also sent in the cookies on log in, expiration date depends
-    on the app, for e-commerce I will leave it at 1 month, but until production I will
-    leave it at 15 minutes for testing. When it expires the user is signed out and cannot
-    access any protected route by the 'verifyAccessToken' middleware   */
-const createRefreshToken = (id) => {
-    return jwt.sign({ id }, process.env.JWT_REFRESH_SECRET, { expiresIn: '15m' })
-}
+const { createAccessToken, createRefreshToken } = require('../../functions/createTokens');
+const { currentUser } = require('../../functions/currentUser')
 
 //Cookies options (in production -> secure:true)
 const refreshOptions = {
@@ -69,35 +57,53 @@ const signoutUser = async (req, res) => {
     res.status(200).json('Signed out')
 }
 
-const getCurrentUser = async (req, res) => {
-    const authHeader = req.headers['authorization']
-    const accessToken = req.cookies.accessToken || authHeader?.split(' ')[1]
+const getAllUsers = async (req, res) => {
+    // in production -> only admins
     try {
-        const Token = jwt.verify(accessToken, process.env.JWT_ACCESS_SECRET)
-        const User = await userModel.findById(Token.id).select('fullname email')
+        const Users = await userModel.find()
+        res.status(200).json(Users)
+    } catch (err) {
+        res.status(400).json({ error: err.message })
+    }
+}
+
+const getCurrentUser = async (req, res) => {
+    try {
+        const User = await currentUser(req, res)
         if (!User) {
-            return res.status(404).json({ error: "InvalidId" });
+            return res.status(403).json({ error: "InvalidId" });
         }
-        /*  If the 'return' header is true, then I want to use this controller
-            as a function in other controllers. And I want it to just return the
-            value, not respond   */
-        return req.headers['return'] !== true ? res.status(200).json(User) : User
+        res.status(200).json(User)
     } catch (error) {
         res.status(400).json({ error: error.message })
     }
 }
 
 const deleteUser = async (req, res) => {
+    const { id } = req.params;
     try {
-        const { id } = req.params;
-        const User = await userModel.findByIdAndDelete(id);
+        //Check current user
+        const User = await currentUser(req, res)
         if (!User) {
-            return res.status(404).json({ error: "InvalidId" });
+            return res.status(403).json({ error: "InvalidId" });
         }
+        //Check user authorization
+        if (User._id.toString() !== id && User.role !== 'admin') {
+            return res.status(403).json({ error: "Forbidden" });
+        }
+        await userModel.findByIdAndDelete(id);
         res.status(200).json("User deleted successfully");
     } catch (err) {
         res.status(400).json({ error: err.message });
     }
 }
 
-module.exports = { postUser, logUser, createAccessToken, signoutUser, getCurrentUser, deleteUser }
+module.exports = {
+    postUser,
+    logUser,
+    createAccessToken,
+    signoutUser,
+    getCurrentUser,
+    deleteUser,
+    getAllUsers
+}
